Removed an unnecessary static_cast.

+++ src/rdb_protocol/func.cc
@@ -123,8 +123,7 @@ counted_t<val_t> func_t::call(const std::vector<counted_t<const datum_t> > &args
                 result);
         } else {
             r_sanity_check(body.has() && source.has() && js_env == NULL);
-            rcheck(args.size() == static_cast<size_t>(argptrs.size())
-                   || argptrs.size() == 0,
+            rcheck(args.size() == argptrs.size() || argptrs.size() == 0,
                    base_exc_t::GENERIC,
                    strprintf("Expected %zd argument(s) but found %zu.",
                              argptrs.size(), args.size()));
